mini game part 3 with proper keyboard text input using import ctype
